Created gluon service and proliferating the use of it where applicable.
