+++ src/gluon/util/project/Project.ts
@@ -1,4 +1,6 @@
+import {HandleCommand, HandlerContext} from "@atomist/automation-client";
 import * as _ from "lodash";
+import {createMenu} from "../shared/GenericMenu";
 
 export function getProjectId(tenant: string, project: string, environment: string): string {
     return `${_.kebabCase(tenant).toLowerCase()}-${_.kebabCase(project).toLowerCase()}-${environment.toLowerCase()}`;
@@ -14,4 +16,21 @@ export function getProjectDisplayName(tenant: string, project: string, environme
     }
 
     return `${tenant} ${project} ${environment.toUpperCase()}`;
+}
+
+export function menuForProjects(ctx: HandlerContext, projects: any[],
+                                command: HandleCommand, message: string = "Please select a project",
+                                projectNameVariable: string = "projectName"): Promise<any> {
+    return createMenu(ctx,
+        projects.map(project => {
+            return {
+                value: project.name,
+                text: project.name,
+            };
+        }),
+        command,
+        message,
+        "Select Project",
+        projectNameVariable,
+    );
 }
